requirements.txt
Django>=3.0.5,<3.1.0
djangorestframework>=3.11.0,<3.12.0
tweepy>=3.8.0,<3.9.0
django-cors-headers>=3.2.1,<3.3.0